requirements.txt
click>=8.1.0
quart>=0.18.0
hypercorn>=0.14.0
playwright>=1.30.0
easyocr>=1.7.0
numpy>=1.22.0,<2.0
markdown>=3.4.0
python-multipart>=0.0.6
aiofiles>=0.8.0
pydantic>=1.10.0
aiohttp>=3.8.0
beautifulsoup4>=4.9.0
google-cloud-tasks>=2.13.0
google-cloud-storage>=2.9.0
google-cloud-firestore>=2.11.0
aioredis>=2.0.1
redis>=4.5.0
anthropic>=0.23.1
fastmcp>=0.5.0
phonenumbers>=8.13.0
email-validator>=2.0.0
google-cloud-ndb>=2.0.0
python-dotenv>=1.0.0
sendgrid>=6.10.0
twilio>=8.0.0
coolname>=2.0.0
